requirements.txt
streamlit 
numpy 
matplotlib 
pillow
